All: Add logging implementation with verbosity
Project had reached a point where there were too many log lines
being thrown around, so I imported some standard crates for
logging with user-specified verbosity.

Verbosity is specified using the -v flag, with the number of
occurrences of the 'v' character mapping to the amount of
verbosity of the logs.

If the user does not enter any verbosity, the program
defaults to Info logging. One additional verbosity
character brings the user to Debug logging, with one
more bringing the user to full Trace logging.