Core: Add a common Image struct for printing out final results
Added a new Image struct and relevant dependency structs to better
represent the final output of an "image" in a format-agnostic manner.

Main method can still use the raw images in their appropriate formats,
of course, but this provides a common way to perform analysis on an
image without needing to support each individual image format in future
code.